Changes of version: 20090501

+++ libewf/libewf_metadata.c
@@ -1988,8 +1988,8 @@ int libewf_handle_get_amount_of_acquiry_errors(
 int libewf_handle_get_acquiry_error(
      libewf_handle_t *handle,
      uint32_t index,
-     off64_t *first_sector,
-     uint32_t *amount_of_sectors,
+     uint64_t *first_sector,
+     uint64_t *amount_of_sectors,
      liberror_error_t **error )
 {
 	static char *function = "libewf_handle_get_acquiry_error";
@@ -2030,8 +2030,8 @@ int libewf_handle_get_acquiry_error(
  */
 int libewf_handle_add_acquiry_error(
      libewf_handle_t *handle,
-     off64_t first_sector,
-     uint32_t amount_of_sectors,
+     uint64_t first_sector,
+     uint64_t amount_of_sectors,
      liberror_error_t **error )
 {
 	static char *function = "libewf_handle_add_acquiry_error";
@@ -2134,8 +2134,8 @@ int libewf_handle_get_amount_of_crc_errors(
 int libewf_handle_get_crc_error(
      libewf_handle_t *handle,
      uint32_t index,
-     off64_t *first_sector,
-     uint32_t *amount_of_sectors,
+     uint64_t *first_sector,
+     uint64_t *amount_of_sectors,
      liberror_error_t **error )
 {
 	libewf_internal_handle_t *internal_handle = NULL;
@@ -2190,8 +2190,8 @@ int libewf_handle_get_crc_error(
  */
 int libewf_handle_add_crc_error(
      libewf_handle_t *handle,
-     off64_t first_sector,
-     uint32_t amount_of_sectors,
+     uint64_t first_sector,
+     uint64_t amount_of_sectors,
      liberror_error_t **error )
 {
 	libewf_internal_handle_t *internal_handle = NULL;
@@ -2297,8 +2297,8 @@ int libewf_handle_get_amount_of_sessions(
 int libewf_handle_get_session(
      libewf_handle_t *handle,
      uint32_t index,
-     off64_t *first_sector,
-     uint32_t *amount_of_sectors,
+     uint64_t *first_sector,
+     uint64_t *amount_of_sectors,
      liberror_error_t **error )
 {
 	static char *function = "libewf_handle_get_session";
@@ -2339,8 +2339,8 @@ int libewf_handle_get_session(
  */
 int libewf_handle_add_session(
      libewf_handle_t *handle,
-     off64_t first_sector,
-     uint32_t amount_of_sectors,
+     uint64_t first_sector,
+     uint64_t amount_of_sectors,
      liberror_error_t **error )
 {
 	static char *function = "libewf_handle_add_session";
